requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
openai==1.54.0
python-dotenv==1.0.0
caldav==1.3.9
icalendar==5.0.11
pytz==2023.3
psycopg2-binary==2.9.9
